refactor: compose capability operation fragments
