+++ src/shared/contracts/operations/campaign-rules.ts
@@ -1,0 +1,20 @@
+import {
+  campaignRulesCommandReceiptInputSchema,
+  campaignRulesSchema,
+  updateCampaignRulesInputSchema
+} from '../campaign-rules.js'
+import { none, read, write } from './registry.js'
+
+export const campaignRulesOperationDefinitions = {
+  'campaignRules.read': read('campaign-rules:read', none, campaignRulesSchema),
+  'campaignRules.update': write(
+    'campaign-rules:update',
+    updateCampaignRulesInputSchema,
+    campaignRulesSchema
+  ),
+  'campaignRules.commandReceipt': read(
+    'campaign-rules:command-receipt',
+    campaignRulesCommandReceiptInputSchema,
+    campaignRulesSchema.nullable()
+  )
+} as const
