+++ src/shared/contracts/operations/encounter.ts
@@ -1,0 +1,13 @@
+import {
+  encounterSelectionEvaluationSchema,
+  evaluateEncounterSelectionInputSchema
+} from '../scene.js'
+import { read } from './registry.js'
+
+export const encounterOperationDefinitions = {
+  'encounter.evaluate': read(
+    'encounter:evaluate',
+    evaluateEncounterSelectionInputSchema,
+    encounterSelectionEvaluationSchema
+  )
+} as const
