+++ src/shared/contracts/operations/npcs.ts
@@ -1,0 +1,47 @@
+import {
+  createWorldNpcInputSchema,
+  deleteWorldNpcInputSchema,
+  updateWorldNpcInputSchema,
+  worldNpcCommandReceiptInputSchema,
+  worldNpcCommandReceiptSchema,
+  worldNpcDeleteReceiptSchema,
+  worldNpcDetailInputSchema,
+  worldNpcDetailProjectionSchema,
+  worldNpcMutationReceiptSchema,
+  worldNpcPageSchema,
+  worldNpcSearchInputSchema
+} from '../world-npc.js'
+import { read, write } from './registry.js'
+
+export const npcsOperationDefinitions = {
+  'npcs.search': read(
+    'npcs:search',
+    worldNpcSearchInputSchema,
+    worldNpcPageSchema
+  ),
+  'npcs.detail': read(
+    'npcs:detail',
+    worldNpcDetailInputSchema,
+    worldNpcDetailProjectionSchema
+  ),
+  'npcs.commandReceipt': read(
+    'npcs:command-receipt',
+    worldNpcCommandReceiptInputSchema,
+    worldNpcCommandReceiptSchema.nullable()
+  ),
+  'npcs.create': write(
+    'npcs:create',
+    createWorldNpcInputSchema,
+    worldNpcMutationReceiptSchema
+  ),
+  'npcs.update': write(
+    'npcs:update',
+    updateWorldNpcInputSchema,
+    worldNpcMutationReceiptSchema
+  ),
+  'npcs.delete': write(
+    'npcs:delete',
+    deleteWorldNpcInputSchema,
+    worldNpcDeleteReceiptSchema
+  )
+} as const
